Refactored engangsstonad and routing
