+++ src/app/bootstrap.js
@@ -1,11 +1,10 @@
 import React from 'react';
 import { render } from 'react-dom';
-import { BrowserRouter as Router, Switch } from 'react-router-dom';
+import { BrowserRouter as Router } from 'react-router-dom';
 import { Provider } from 'react-redux';
 
+import EngangsstonadIndex from 'components/engangsstonad/EngangsstonadIndex';
 import IntlProvider from './../app/components/intl/IntlProvider';
-import renderTopLevelRoutes from './util/routing';
-import routeConfig from './util/routing/routes.config';
 import store from './redux';
 
 const root = document.getElementById('app');
@@ -14,7 +13,7 @@ render(
 		<IntlProvider>
 			<Router>
 				<div className="app">
-					<Switch>{renderTopLevelRoutes(routeConfig)}</Switch>
+					<EngangsstonadIndex />
 				</div>
 			</Router>
 		</IntlProvider>
