--- src/app/components/engangsstonad/engangsstonadContainer.less
@@ -1,9 +1,0 @@
-.linkIndicatorWrapper {
-	position: relative;
-	margin-bottom: 16px;
-
-	&__link {
-		position: absolute;
-		bottom: 0.35rem;
-	}
-}
