+++ src/app/util/modules/getModulesFromContext.js
@@ -1,9 +1,9 @@
-const getModulesFromContext = (context) => {
-	const modules = [];
-	context.keys().forEach((ref) => {
-		modules.push(context(ref).default);
-	});
-	return modules;
-};
+// const getModulesFromContext = (context) => {
+// 	const modules = [];
+// 	context.keys().forEach((ref) => {
+// 		modules.push(context(ref).default);
+// 	});
+// 	return modules;
+// };
 
-export default getModulesFromContext;
+// export default getModulesFromContext;
