+++ src/app/util/routing/index.js
@@ -1,8 +1,8 @@
-import React from 'react';
+// import React from 'react';
 
-import RouteWithSubRoutes from './route.component';
+// import RouteWithSubRoutes from './route.component';
 
-export default (routeConfig) =>
-	routeConfig.map((route) => (
-		<RouteWithSubRoutes {...route} key={route.path || '404'} />
-	));
+// export default (routeConfig) =>
+// 	routeConfig.map((route) => (
+// 		<RouteWithSubRoutes {...route} key={route.path || '404'} />
+// 	));
