+++ src/app/util/routing/route.component.js
@@ -1,15 +1,15 @@
-import React from 'react';
-import { Route as ReactRoute } from 'react-router-dom';
+// import React from 'react';
+// import { Route as ReactRoute } from 'react-router-dom';
 
-const RouteWithSubRoutes = (route) => (
-	<ReactRoute
-		path={route.path}
-		exact={route.exact}
-		render={(props) => (
-			<route.component {...props} routes={route.routes} key={route.path} />
-		)}
-		key={route.path}
-	/>
-);
+// const RouteWithSubRoutes = (route) => (
+// 	<ReactRoute
+// 		path={route.path}
+// 		exact={route.exact}
+// 		render={(props) => (
+// 			<route.component {...props} routes={route.routes} key={route.path} />
+// 		)}
+// 		key={route.path}
+// 	/>
+// );
 
-export default RouteWithSubRoutes;
+// export default RouteWithSubRoutes;
